Attempt to reconstruct corrupt tarsplit files
1.10 and 1.10.1 had a bug that could create truncated tarsplit files
when migrating layers from older versions. If an attempt to serialize a
layer to tar fails because of an "unexpected EOF" error, try creating
the tarsplit file based on the contents of the layer in the graph
driver. If this results in a matching DiffID, replace the corrupt
tarsplit file with the fixed one. This will allow future docker push
retries and docker save attempts to succeed.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>