Synchronous service create and service update
Change "service create" and "service update" to wait until the creation
or update finishes, by default. Show progress bars for the overall
operation and for each individual task (when there are a small enough
number of tasks).

Using "-d" or "--detach" will cause the old, asynchronous behavior.

Using "-q" or "--quiet" will make the commands still wait for the
creation or update to complete, but without outputting progress
information.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>